chore: Added Glint signature to <NavigationMenu> component
