--- tests/dummy/app/components/navigation-menu.d.ts
@@ -1,9 +1,0 @@
-type MenuItem = {
-  label: string;
-  route: string;
-};
-
-export interface NavigationMenuComponentArgs {
-  menuItems: Array<MenuItem>;
-  name?: string;
-}
