+++ tests/dummy/app/components/navigation-menu.hbs
@@ -1,4 +1,3 @@
-{{! @glint-nocheck: not typesafe yet }}
 <nav aria-label={{@name}} data-test-nav={{@name}}>
   <ul local-class="list">
     {{#each @menuItems as |menuItem|}}
