+++ tests/dummy/app/components/navigation-menu.ts
@@ -1,0 +1,24 @@
+import templateOnlyComponent from '@ember/component/template-only';
+
+type MenuItem = {
+  label: string;
+  route: string;
+};
+
+export interface NavigationMenuComponentSignature {
+  Args: {
+    menuItems: Array<MenuItem>;
+    name?: string;
+  };
+}
+
+const NavigationMenuComponent =
+  templateOnlyComponent<NavigationMenuComponentSignature>();
+
+export default NavigationMenuComponent;
+
+declare module '@glint/environment-ember-loose/registry' {
+  export default interface Registry {
+    NavigationMenu: typeof NavigationMenuComponent;
+  }
+}
